Remove default exports from workbox-routing
