--- packages/workbox-routing/_public.mjs
@@ -1,22 +1,0 @@
-/*
-  Copyright 2018 Google LLC
-
-  Use of this source code is governed by an MIT-style
-  license that can be found in the LICENSE file or at
-  https://opensource.org/licenses/MIT.
-*/
-
-import {DefaultRouter} from './DefaultRouter.mjs';
-import {RegExpRoute} from './RegExpRoute.mjs';
-import {Route} from './Route.mjs';
-import {Router} from './Router.mjs';
-import {NavigationRoute} from './NavigationRoute.mjs';
-import './_version.mjs';
-
-export {
-  DefaultRouter,
-  RegExpRoute,
-  Route,
-  Router,
-  NavigationRoute,
-};
